Simplify home page
