+++ src/components/home/index.js
@@ -1,0 +1,3 @@
+export { default as Hero } from './Hero';
+export { default as FooterCTA } from './FooterCTA';
+export * as features from './FeatureBlocks';
